Moved LabeledInput to features/form/components/fragments folder and fixed import path
