+++ app/client/src/common/components/index.js
@@ -4,7 +4,7 @@
 
 import Divider from './Divider'
 import Icon from './Icon'
-import LabeledInput, { Label, Input } from './LabeledInput'
+import LabeledInput, { Label, Input } from '../../features/form/components/fragments/LabeledInput'
 import Button from './Button'
 import RoundButton from './RoundButton'
 import ScrollToTop from './ScrollToTop'
