+++ app/client/src/features/form/components/fragments/Job.js
@@ -6,12 +6,12 @@ import React from 'react'
 import styled from 'styled-components'
 import {
   Divider,
-  LabeledInput,
   Label,
   Input,
   RoundButton,
   Icon
 } from '../../../../common/components'
+import LabeledInput from '../fragments/LabeledInput'
 
 const Row = styled.div`
   display: flex;
