+++ app/client/src/features/form/components/sections/Education.js
@@ -5,7 +5,8 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import Section from './Section'
-import { Button, LabeledInput, Divider } from '../../../../common/components'
+import { Button, Divider } from '../../../../common/components'
+import LabeledInput from '../fragments/LabeledInput'
 import { School } from '..'
 import { addSchool, removeSchool } from '../../actions'
 import type { FormValues } from '../../types'
